Update README.md
I wanted to practice to see how far I have gone with R programming so I saw a dataset for London bike trips from 2015 to 2017 and it was gotten from the public dataset from big query. So I used Rstudio to analyse this data to find the station with the highest rentals in London from 2015 to 2017 and I also found out the station with the highest rental in each year and I found out that Hyde Corner Park, Hyde Park was the station with the highest rental in each year from 2015 to 2017. Thou their sales reduced in each year especially 2017 so I also conducted an analysis to find out why and I found out that Hyde corner park made more sales from people returning back to the station as their final destination and each year the number of people returning reduced especially the year 2017.  A code of how I analysed this in Rstudio is being placed above. Please you could view it to understand how I came up with my findings. Thank you